+++ cart/Changelog.md
@@ -58,3 +58,11 @@
 # 10. February 2020
 * Add `additionalData` to `AddRequest` 
   * Breaking change: Update helper/builder function `BuildAddRequest`
+* Breaking Change to `EventPublisher` interface, `PublishChangedQtyInCartEvent` and `PublishAddToCartEvent` now
+include a cart as a parameter
+* Breaking Change to behaviour of `AddToCartEvent` and `ChangedQtyInCartEvent`, they are now thrown after
+the cart has been adjusted and written back to cache
+* Events deferred from `ModifyBehaviour` are dispatched before `AddToCartEvent` and `ChangedQtyInCartEvent`
+* The `AddToCartEvent` includes the current cart (with added product)
+* The `ChangedQtyInCartEvent` includes the current cart (with updated quantities)
+* Add Whitebox Test `TestCartService_CartInEvent` to check `AddToCartEvent`